updated 21/07/22

+++ client/src/pages/admin/PagesAdmin.js
@@ -24,12 +24,18 @@ const PagesAdmin = () => {
   const dispatch = useDispatch();
 
   async function handleDelete(slug) {
+    let token =  localStorage.getItem("auth");
+  if (token) {
     let result = await fetch(`${BASE_URL}/admin/page/${slug}`, {
       method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
     });
     result = await result.json();
     dispatch(getAllPage());
   }
+}
 
   useEffect(() => {
     setPageList(pages);
